Temporary work-around in order to get BT to compile.
error on line 127:
killer = proj.getShooter().getType().getName();

changed to:
killer = proj.getShotter().toString();

as a temporary work-around. My primary goal was to get BattleArena to
compile, but in order to do that, I had to build each dependency for it...
BattleTracker being ones of those dependencies.

Will have to investigate later exactly what the problem with this line is.